--- services-api/src/main/java/io/scalecube/services/transport/api/experimental/ServiceTransportProvider.java
@@ -1,8 +1,0 @@
-package io.scalecube.services.transport.api.experimental;
-
-public interface ServiceTransportProvider {
-
-  ClientTransportFactory provideClientTransportFactory();
-
-  ServerTransport provideServerTransport();
-}
