+++ services-bytebuf-codec/src/main/java/io/scalecube/services/transport/api/BinaryHeadersCodec.java
@@ -13,13 +13,22 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
 
+/**
+ * Simple binary codec for headers service message.
+ */
 public class BinaryHeadersCodec implements HeadersCodec {
 
+  /**
+   * {@inheritDoc}
+   */
   @Override
   public String contentType() {
     return "application/octet-stream";
   }
 
+  /**
+   * {@inheritDoc}
+   */
   @Override
   public void encode(OutputStream stream, Map<String, String> headers) throws IOException {
     if (headers.isEmpty()) {
@@ -53,6 +62,9 @@ public class BinaryHeadersCodec implements HeadersCodec {
     }
   }
 
+  /**
+   * {@inheritDoc}
+   */
   @Override
   public Map<String, String> decode(InputStream stream) throws IOException {
     if (stream.available() < 1) {
