+++ services-examples-parent/services-examples/src/main/java/io/scalecube/services/examples/ServiceTransports.java
@@ -1,7 +1,7 @@
 package io.scalecube.services.examples;
 
 import io.scalecube.services.Microservices.ServiceTransportBootstrap;
-import io.scalecube.services.transport.rsocket.experimental.builder.RSocketByNettyTcp;
+import io.scalecube.services.transport.rsocket.builder.RSocketByNettyTcp;
 
 public class ServiceTransports {
