+++ services-transport-parent/services-transport-rsocket/src/main/java/io/scalecube/services/transport/rsocket/RSocketServiceAcceptor.java
@@ -37,6 +37,9 @@ public class RSocketServiceAcceptor implements SocketAcceptor {
     this.methodRegistry = methodRegistry;
   }
 
+  /**
+   * {@inheritDoc}
+   */
   @Override
   public Mono<RSocket> accept(ConnectionSetupPayload setup, RSocket socket) {
     LOGGER.info("Accepted rSocket: {}, connectionSetup: {}", socket, setup);
