--- services-transport-parent/services-transport-rsocket/src/main/java/io/scalecube/services/transport/rsocket/experimental/RSocketClientTransportFactory.java
@@ -1,9 +1,0 @@
-package io.scalecube.services.transport.rsocket.experimental;
-
-import io.rsocket.transport.ClientTransport;
-import io.scalecube.net.Address;
-
-public interface RSocketClientTransportFactory {
-
-  ClientTransport createClient(Address address);
-}
