--- services-transport-parent/services-transport-rsocket/src/main/java/io/scalecube/services/transport/rsocket/experimental/RSocketServerTransportFactory.java
@@ -1,15 +1,0 @@
-package io.scalecube.services.transport.rsocket.experimental;
-
-import io.rsocket.Closeable;
-import io.rsocket.transport.ServerTransport;
-import io.scalecube.net.Address;
-
-public interface RSocketServerTransportFactory {
-
-  ServerTransport<Server> createServer(Address address);
-
-  interface Server extends Closeable {
-
-    Address address();
-  }
-}
