+++ services-transport-parent/services-transport-rsocket/src/main/java/io/scalecube/services/transport/rsocket/tcp/ExtendedNioEventLoopGroup.java
@@ -1,23 +1,20 @@
-package io.scalecube.services.transport.rsocket.experimental.tcp;
+package io.scalecube.services.transport.rsocket.tcp;
 
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
 import io.netty.channel.ChannelPromise;
-import io.netty.channel.DefaultSelectStrategyFactory;
 import io.netty.channel.EventLoop;
-import io.netty.channel.EventLoopGroup;
-import io.netty.channel.MultithreadEventLoopGroup;
 import io.netty.channel.SelectStrategy;
 import io.netty.channel.SelectStrategyFactory;
+import io.netty.channel.nio.NioEventLoopGroup;
 import io.netty.util.concurrent.DefaultThreadFactory;
 import io.netty.util.concurrent.RejectedExecutionHandler;
-import io.netty.util.concurrent.RejectedExecutionHandlers;
 import java.lang.reflect.Constructor;
+import java.nio.channels.spi.SelectorProvider;
 import java.util.concurrent.Executor;
 import java.util.concurrent.ThreadFactory;
 
-public class ExtendedEpollEventLoopGroup extends MultithreadEventLoopGroup
-    implements ExtendedEventLoopGroup {
+public class ExtendedNioEventLoopGroup extends NioEventLoopGroup implements ExtendedEventLoopGroup {
 
   private static final ThreadFactory WORKER_THREAD_FACTORY =
       new DefaultThreadFactory("rsocket-worker", true);
@@ -27,24 +24,19 @@ public class ExtendedEpollEventLoopGroup extends MultithreadEventLoopGroup
    *
    * @param numOfThreads number of worker threads
    */
-  public ExtendedEpollEventLoopGroup(int numOfThreads) {
-    super(
-        numOfThreads,
-        WORKER_THREAD_FACTORY,
-        0,
-        DefaultSelectStrategyFactory.INSTANCE,
-        RejectedExecutionHandlers.reject());
+  public ExtendedNioEventLoopGroup(int numOfThreads) {
+    super(numOfThreads, WORKER_THREAD_FACTORY);
   }
 
   @Override
   protected EventLoop newChild(Executor executor, Object... args) throws Exception {
-    Class<?> eventLoopClass = Class.forName("io.netty.channel.epoll.EpollEventLoop");
+    Class<?> eventLoopClass = Class.forName("io.netty.channel.nio.NioEventLoop");
 
     Constructor<?> constructor =
         eventLoopClass.getDeclaredConstructor(
-            EventLoopGroup.class,
+            NioEventLoopGroup.class,
             Executor.class,
-            int.class,
+            SelectorProvider.class,
             SelectStrategy.class,
             RejectedExecutionHandler.class);
 
@@ -54,7 +46,7 @@ public class ExtendedEpollEventLoopGroup extends MultithreadEventLoopGroup
         constructor.newInstance(
             this,
             executor,
-            (Integer) args[0],
+            (SelectorProvider) args[0],
             ((SelectStrategyFactory) args[1]).newSelectStrategy(),
             (RejectedExecutionHandler) args[2]);
   }
