+++ services/src/test/java/io/scalecube/services/transport/ServiceTransports.java
@@ -1,7 +1,7 @@
 package io.scalecube.services.transport;
 
 import io.scalecube.services.Microservices.ServiceTransportBootstrap;
-import io.scalecube.services.transport.rsocket.experimental.builder.RSocketByNettyTcp;
+import io.scalecube.services.transport.rsocket.builder.RSocketByNettyTcp;
 
 public class ServiceTransports {
